64 bit cleanup
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@13105 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 1f8d7c21b88600f4c99f398e61ed45c0dc0478ae

+++ vision/calonder_descriptor/src/py.cpp
@@ -231,7 +231,7 @@ PyObject *train(PyObject *self, PyObject *args)
                           &x, &y, &kp, &num_trees, &depth, &views,
                           &dimension, &num_quant_bits))
       return NULL;
-    dimension = std::min(dimension, PyList_Size(kp));
+    dimension = std::min(dimension, (int)PyList_Size(kp));
     input = cvCreateImageHeader(cvSize(x, y), IPL_DEPTH_8U, 1);
     cvSetData(input, imgdata, x);
   }
